Reduced compiler warnings to a minimal set
There are 5 remaining compiler warnings, three of which are around code
that has comments stating it is not currently used, which implies that
we may want to use it in the future.

The remaining two are in the Events class, which we should remove at
some point as we transition to RxJava plumbing internally.  The Events
code is also a bit hairy, so it looked best to leave this as-is for the
moment.