Simplify check scheduling by introduction of explicit Schedule
RemoteBFTCheckSchedule defines when checks are to be run according to some arbitrary schedule. To facilitate this, the RemoteBFTCheck interface has been amended to represent a single check per call rather than a continuous source of check results. This serves to further streamline the check pipeline and makes testing of checks more straightforward.